User Authorization and Authentication Done and Route Protection done using JWT

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 import asyncHandler from "express-async-handler";
 import Goal from "../models/goalModel.js";
+import User from "../models/userModel.js";
 
 // @desc GET goals
 // @route GET /api/goals
 // @access private
 export const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).send(goals);
 });
@@ -20,6 +21,7 @@ export const setGoal = asyncHandler(async (req, res) => {
   } else {
     const goal = await Goal.create({
       text: req.body.text,
+      user: req.user.id,
     });
     res.status(201).json(goal);
   }
@@ -34,12 +36,24 @@ export const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal Not Found!");
-  } else {
-    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-    });
-    res.status(200).json(updatedGoal);
   }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User Not Found!");
+  }
+
+  if (goal.user.toString() !== user) {
+    res.status(401);
+    throw new Error("User not Authorized!");
+  }
+
+  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+  });
+  res.status(200).json(updatedGoal);
 });
 
 // @desc delete goal
@@ -50,8 +64,20 @@ export const deleteGoal = asyncHandler(async (req, res) => {
 
   if (!goal) {
     throw new Error("Goal Not Found!");
-  } else {
-    const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
-    res.status(201).json(deletedGoal);
   }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User Not Found!");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not Authorized!");
+  }
+
+  const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
+  res.status(201).json(deletedGoal);
 });
